refactor(issue-11): modularize agent nodes, parallelize bard/accountant/scribe
